Fix out-of-range limit

+++ src/subcommand/wallet/transactions.rs
@@ -10,7 +10,7 @@ impl Transactions {
   pub(crate) fn run(self, options: Options) -> Result {
     for tx in options.bitcoin_rpc_client()?.list_transactions(
       None,
-      Some(self.limit.unwrap_or(usize::MAX)),
+      Some(self.limit.unwrap_or(u16::MAX as usize)),
       None,
       None,
     )? {
